Enable PKG_CONFIG and OS to be overriden on the command-line
On a cross-compiled build, there is need for an invocation of pkg-config which
is prefixed with the host target. (eg. i686-w64-mingw32-pkg-config)
This allows to pass PKG_CONFIG as a variable to make, using the default if
none is given.

Also, the variable OS allows to override the system identified by uname, to
handle a cross-compiling situation for Mac OS targets.